댓글 작성 처리. 리스팅. 카운트 기능 구현

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -1,59 +1,65 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <c:set var="pageTitle" value="ARTICLE DETAIL" />
 <%@ include file="../common/head.jsp"%>
 
 <script>
 	const params = {};
-	params.id = parseInt('${param.id}')
-	function ArticlDetail__increaseHitCount() {
+	params.id = parseInt('${param.id}');
+	
+	function ArticleDetail__increaseHitCount() {
+		
 		const localStorageKey = 'article__' + params.id + '__alreadyView';
+		
 		if (localStorage.getItem(localStorageKey)) {
 			return;
 		}
 		localStorage.setItem(localStorageKey, true);
+		
 		$.get('doIncreaseHitCountRd', {
 			id : params.id,
 			ajaxMode : 'Y'
-		}, function(data) {
+		}, function(data){
 			$('.article-detail__hit-count').empty().html(data.data1);
 		}, 'json');
 	}
+	
 	function ReactionPoint__getReactionPoint() {
 		
 		$.get('../reactionPoint/getReactionPoint', {
 			id : params.id,
+			relTypeCode : 'article',
 			ajaxMode : 'Y'
 		}, function(data){
 			if(data.data1.sumReactionPoint > 0){
 				let goodBtn = $('#goodBtn'); 
 				goodBtn.removeClass('btn-outline');
-// 				goodBtn.prop('href', '취소되는 요청으로')
 				goodBtn.prop('href', '../reactionPoint/delReactionPoint?id=${article.id}&relTypeCode=article&point=1')
 			}else if(data.data1.sumReactionPoint < 0){
 				let badBtn = $('#badBtn');
 				badBtn.removeClass('btn-outline');
-// 				badBtn.prop('href', '취소되는 요청으로')
 				badBtn.prop('href', '../reactionPoint/delReactionPoint?id=${article.id}&relTypeCode=article&point=-1')
 			}
 		}, 'json');
 	}
 	
-	$(function() {
-		/* 실전 코드*/
-		/* ArticlDetail__increaseHitCount(); */
+	$(function(){
+// 		실전코드
+		ArticleDetail__increaseHitCount();
 		ReactionPoint__getReactionPoint();
-		/* 연습코드 */
-		setTimeout(ArticlDetail__increaseHitCount, 3000)
+		
+// 		연습코드
+// 		setTimeout(ArticleDetail__increaseHitCount, 2000);
 	})
+	
 </script>
+
 <section class="mt-8 text-xl">
 	<div class="container mx-auto px-3 pb-5 border-bottom-line">
 		<div class="table-box-type-1">
 			<table>
 				<colgroup>
-					<col width="200" />
+					<col width="200"/>
 				</colgroup>
 
 				<tbody>
@@ -71,8 +77,7 @@
 					</tr>
 					<tr>
 						<th>조회수</th>
-						<td><span class="badge article-detail__hit-count">
-								${article.hitCount} </span></td>
+						<td><span class="badge article-detail__hit-count">${article.hitCount}</span></td>
 					</tr>
 					<tr>
 						<th>작성자</th>
@@ -80,17 +85,20 @@
 					</tr>
 					<tr>
 						<th>추천</th>
-						<td><c:if test="${rq.getLoginedMemberId() == 0 }">
+						<td>
+							<c:if test="${rq.getLoginedMemberId() == 0 }">
 								<span class="badge">좋아요 : ${article.goodReactionPoint}개</span>
 								<br />
 								<span class="badge">싫어요 : ${article.badReactionPoint * -1}개</span>
-							</c:if> <c:if test="${rq.getLoginedMemberId() != 0 }">
+							</c:if>
+							<c:if test="${rq.getLoginedMemberId() != 0 }">
 								<a id="goodBtn" class="btn btn-xs btn-outline" href="../reactionPoint/doReactionPoint?id=${article.id }&relTypeCode=article&point=1">좋아요 👍</a>
 								<span class="badge">좋아요 : ${article.goodReactionPoint}개</span>
 								<br />
 								<a id="badBtn" class="btn btn-xs btn-outline" href="../reactionPoint/doReactionPoint?id=${article.id }&relTypeCode=article&point=-1">싫어요 👎</a>
 								<span class="badge">싫어요 : ${article.badReactionPoint * -1}개</span>
-							</c:if></td>
+							</c:if>
+						</td>
 					</tr>
 					<tr>
 						<th>제목</th>
@@ -104,65 +112,54 @@
 			</table>
 		</div>
 		<div class="btns mt-2">
-			<button class="btn-text-link btn btn-active btn-ghost" type="button"
-				onclick="history.back();">뒤로가기</button>
+			<button class="btn-text-link btn btn-active btn-ghost" type="button" onclick="history.back();">뒤로가기</button>
 			<c:if test="${article.actorCanChangeData }">
-				<a class="btn-text-link btn btn-active btn-ghost"
-					href="modify?id=${article.id }">수정</a>
-				<a class="btn-text-link btn btn-active btn-ghost"
-					onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;"
-					href="doDelete?id=${article.id }">삭제</a>
+				<a class="btn-text-link btn btn-active btn-ghost" href="modify?id=${article.id }">수정</a>
+				<a class="btn-text-link btn btn-active btn-ghost" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;" href="doDelete?id=${article.id }">삭제</a>
 			</c:if>
 		</div>
 	</div>
 </section>
+
 <script>
-	function ReplyWrite__submitForm(form){
+	function ReplyWrite__submitForm(form) {
+		
 		form.body.value = form.body.value.trim();
 		
-		if(form.body.value.length<2){
-			alert('2글자 이상 입력해 주세요');
+		if(form.body.value.length < 2) {
+			alert('2글자 이상 입력해주세요');
 			form.body.focus();
 			return;
-			
 		}
+		
 		form.submit();
-		
 	}
-	
 </script>
+
 <section class="mt-5 text-xl">
-	<div class="container mx-auto px-3 pb-5 border-bottom-line">
-		<h2>댓글</h2>
-		
-		<!-- 		반복문 돌려서 list처리 여기서부터 -->
-		<div class="py-2 pl-16 border-bottom-line text-base">
-			<div class="font-semibold"><span>작성자</span></div>
-			<div><span>내용</span></div>
-			<div class="text-sm text-gray-400"><span>날짜</span></div>
-		</div>
-<!-- 		여기까지 -->
-		
-		
-		<form action="../reply/doWrite" method="POST" onsubmit="ReplyWrite__submitForm(this); return false">
-			<input type="hidden" name="relTypeCode" value="article" />
-			<input type="hidden" name="relId" value="${article.id }" />
-			<div class="mt-4 p-4 border rounded-lg border-gray-400 text-base">
-				<div class="mb-2"><span>nickname</span></div>
-				<textarea class="textarea textarea-borderd w-full"name="body"rows="2" placeholder="글을 남겨주세요"></textarea>
-				<div class="flex justify-end">
-					<button class="btn btn-active btn-ghost btn-sm">등록</button>
-				</div>
-				
+	<div class="container mx-auto px-3">
+		<h2>댓글<span class="text-base">(${replies.size() }개)</span></h2>
+
+		<c:forEach var="reply" items="${replies}">
+			<div class="py-2 pl-16 border-bottom-line text-base">
+				<div class="font-semibold"><span>${reply.writerName }</span></div>
+				<div><span>${reply.body }</span></div>
+				<div class="text-sm text-gray-400"><span>${reply.updateDate }</span></div>
 			</div>
-		</form>	
+		</c:forEach>
+		
+		<c:if test="${rq.getLoginedMemberId() != 0 }">
+			<form action="../reply/doWrite" method="POST" onsubmit="ReplyWrite__submitForm(this); return false;">
+				<input type="hidden" name="relTypeCode" value="article" />
+				<input type="hidden" name="relId" value="${article.id }" />
+				<div class="mt-4 p-4 rounded-lg border border-gray-400 text-base">
+					<div class="mb-2"><span>현재 로그인한 회원 닉네임</span></div>
+					<textarea class="textarea textarea-bordered w-full" name="body" rows="2" placeholder="댓글을 남겨보세요"></textarea>
+					<div class="flex justify-end"><button class="btn btn-active btn-ghost btn-sm">등록</button></div>
+				</div>
+			</form>
+		</c:if>
 	</div>
-
 </section>
-
-
-
-
-
 
 <%@ include file="../common/foot.jsp"%>
